Asset pricing data from Kenneth R. French's website

A dataset with 329 observations of five factors from North America and Europe.

five

Format

A data frame with 329 rows and 12 columns

Mkt.RF_na

Excess market return in North America

SMB_na

Size factor in North America

HML_na

Value factor in North America

RMW_na

Profitability factor in North America

CMA_na

Investment factor in North America

SMALL.LoBM_na

Return on a portofio of stocks with small market cap and low B/M ratio in North America

Mkt.RF_eu

Excess market return in Europe

SMB_eu

Size factor in Europe

HML_eu

Value factor in Europe

RMW_eu

Profitability factor in Europe

CMA_eu

Investment factor in Europe

SMALL.LoBM_eu

Return on a portofio of stocks with small market cap and low B/M ratio in Europe

Source

subset of data used in Huang (2018)